# server/routes/api-v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Legal Public API",
    "version": "1.0.0",
    "description": "Programmatic access to the AI Legal dataset: global tracker of AI-related lawsuits, regulations, and legal use cases. All responses JSON. No auth required; free and unmetered during the preview period. Identify your app with a descriptive User-Agent so we can reach you if anything breaks.",
    "contact": { "name": "AI Legal", "url": "https://ailegal.co.za" },
    "license": { "name": "CC-BY-4.0", "url": "https://creativecommons.org/licenses/by/4.0/" }
  },
  "servers": [
    { "url": "https://ailegal.co.za/api/v1", "description": "Production" },
    { "url": "http://localhost:3001/api/v1", "description": "Local development" }
  ],
...
  "paths": {
...
    "/lawsuits/{id}": {
      "get": {
        "tags": ["lawsuits"],
        "summary": "Get one lawsuit",
        "description": "Full lawsuit including events timeline, AI-generated insights, and source mentions.",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": {
          "200": { "description": "Lawsuit detail.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LawsuitDetail" } } } },
          "404": { "description": "Not found.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/regulations/{id}": {
      "get": {
        "tags": ["regulations"],
        "summary": "Get one regulation",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": {
          "200": { "description": "Regulation detail.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegulationDetail" } } } },
          "404": { "description": "Not found." }
        }
      }
    },
...
    "/usecases/{id}": {
      "get": {
        "tags": ["usecases"],
        "summary": "Get one use case",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": {
          "200": { "description": "Use case detail.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UseCase" } } } },
          "404": { "description": "Not found." }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
      "LawsuitStatus": { "type": "string", "enum": ["active", "appealing", "settled", "dismissed", "decided"] },
      "LawsuitType":   { "type": "string", "enum": ["copyright", "privacy", "defamation", "labour", "contract", "other"] },
      "RegulationStatus": { "type": "string", "enum": ["proposed", "draft", "consultation", "enacted", "in_force", "partial_force", "amended", "repealed", "superseded"] },
      "RegulationType":   { "type": "string", "enum": ["statute", "regulation", "directive", "guidance", "executive_order", "standard", "voluntary_code", "court_ruling"] },
      "EventType": { "type": "string", "enum": ["filing", "hearing", "ruling", "settlement", "dismissal", "decision", "appeal", "amendment", "update", "enacted", "took_effect", "enforcement_action", "guidance_issued", "proposed", "amended", "repealed"] },
      "LawsuitListItem": {
        "type": "object",
        "properties": {
          "id":             { "type": "string", "format": "uuid" },
          "case_name":      { "type": "string", "example": "Bartz v. Anthropic" },
          "plaintiffs":     { "type": "array", "items": { "type": "string" } },
          "defendants":     { "type": "array", "items": { "type": "string" } },
          "court":          { "type": "string", "nullable": true, "example": "US District Court" },
          "judge":          { "type": "string", "nullable": true },
          "jurisdiction":   { "type": "string", "example": "US Federal" },
          "district":       { "type": "string", "nullable": true },
          "circuit":        { "type": "string", "nullable": true },
          "status":         { "$ref": "#/components/schemas/LawsuitStatus" },
          "case_type":      { "$ref": "#/components/schemas/LawsuitType" },
          "key_issues":     { "type": "array", "items": { "type": "string" } },
          "filing_date":    { "type": "string", "format": "date", "nullable": true },
          "last_update":    { "type": "string", "format": "date", "nullable": true },
          "next_deadline":  { "type": "string", "format": "date", "nullable": true },
          "next_deadline_notes": { "type": "string", "nullable": true },
          "outcome":        { "type": "string", "nullable": true },
          "settlement_amount": { "type": "string", "nullable": true },
          "case_url":       { "type": "string", "format": "uri", "nullable": true },
          "source_url":     { "type": "string", "format": "uri", "nullable": true },
          "source_urls":    { "type": "array", "items": { "type": "string", "format": "uri" } },
          "summary":        { "type": "string", "nullable": true },
          "detailed_analysis": { "type": "string", "nullable": true },
          "tags":           { "type": "array", "items": { "type": "string" } },
          "updated_at":     { "type": "string", "format": "date-time" },
          "latest_event_date":        { "type": "string", "format": "date", "nullable": true },
          "latest_event_type":        { "$ref": "#/components/schemas/EventType" },
          "latest_event_title":       { "type": "string", "nullable": true },
          "latest_event_description": { "type": "string", "nullable": true }
        }
      },
      "LawsuitDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/LawsuitListItem" },
          {
            "type": "object",
            "properties": {
              "events":   { "type": "array", "items": { "$ref": "#/components/schemas/Event" } },
              "insights": { "type": "array", "items": { "$ref": "#/components/schemas/Insight" } },
              "mentions": { "type": "array", "items": { "$ref": "#/components/schemas/Mention" } }
            }
          }
        ]
      },
      "RegulationListItem": {
        "type": "object",
        "properties": {
          "id":                     { "type": "string", "format": "uuid" },
          "regulation_name":        { "type": "string", "example": "Regulation (EU) 2024/1689 — Artificial Intelligence Act" },
          "short_name":             { "type": "string", "nullable": true, "example": "EU AI Act" },
          "jurisdiction":           { "type": "string" },
          "regulator":              { "type": "string", "nullable": true },
          "status":                 { "$ref": "#/components/schemas/RegulationStatus" },
          "regulation_type":        { "$ref": "#/components/schemas/RegulationType" },
          "scope":                  { "type": "array", "items": { "type": "string" } },
          "affected_sectors":       { "type": "array", "items": { "type": "string" } },
          "proposed_date":          { "type": "string", "format": "date", "nullable": true },
          "enacted_date":           { "type": "string", "format": "date", "nullable": true },
          "effective_date":         { "type": "string", "format": "date", "nullable": true },
          "enforcement_date":       { "type": "string", "format": "date", "nullable": true },
          "next_milestone":         { "type": "string", "format": "date", "nullable": true },
          "next_milestone_notes":   { "type": "string", "nullable": true },
          "key_provisions":         { "type": "array", "items": { "type": "string" } },
          "penalties":              { "type": "string", "nullable": true },
          "extraterritorial_scope": { "type": "string", "nullable": true },
          "official_url":           { "type": "string", "format": "uri", "nullable": true },
          "source_url":             { "type": "string", "format": "uri", "nullable": true },
          "source_urls":            { "type": "array", "items": { "type": "string", "format": "uri" } },
          "summary":                { "type": "string", "nullable": true },
          "detailed_analysis":      { "type": "string", "nullable": true },
          "tags":                   { "type": "array", "items": { "type": "string" } },
          "updated_at":             { "type": "string", "format": "date-time" },
          "latest_event_date":      { "type": "string", "format": "date", "nullable": true },
          "latest_event_type":      { "$ref": "#/components/schemas/EventType" },
          "latest_event_title":     { "type": "string", "nullable": true },
          "latest_event_description": { "type": "string", "nullable": true }
        }
      },
      "RegulationDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/RegulationListItem" },
          {
            "type": "object",
            "properties": {
              "events":   { "type": "array", "items": { "$ref": "#/components/schemas/Event" } },
              "insights": { "type": "array", "items": { "$ref": "#/components/schemas/Insight" } },
              "mentions": { "type": "array", "items": { "$ref": "#/components/schemas/Mention" } }
            }
          }
        ]
      },
      "UseCase": {
        "type": "object",
        "properties": {
          "id":                { "type": "string", "format": "uuid" },
          "firm_name":         { "type": "string", "example": "Allen & Overy" },
          "firm_type":         { "type": "string", "enum": ["biglaw", "boutique", "solo", "inhouse", "government", "nonprofit", "legaltech", "other"] },
          "jurisdiction":      { "type": "string" },
          "use_case_title":    { "type": "string" },
          "summary":           { "type": "string" },
          "tools_used":        { "type": "array", "items": { "type": "string" } },
          "categories":        { "type": "array", "items": { "type": "string" } },
          "outcome":           { "type": "string", "nullable": true },
          "quantified_impact": { "type": "string", "nullable": true },
          "source_url":        { "type": "string", "format": "uri", "nullable": true },
          "source_urls":       { "type": "array", "items": { "type": "string", "format": "uri" } },
          "source_name":       { "type": "string", "nullable": true },
          "author":            { "type": "string", "nullable": true },
          "published_at":      { "type": "string", "format": "date", "nullable": true },
          "tags":              { "type": "array", "items": { "type": "string" } },
          "updated_at":        { "type": "string", "format": "date-time" }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id":              { "type": "string", "format": "uuid" },
          "event_date":      { "type": "string", "format": "date", "nullable": true },
          "event_type":      { "$ref": "#/components/schemas/EventType" },
          "title":           { "type": "string", "nullable": true },
          "description":     { "type": "string", "nullable": true },
          "source_url":      { "type": "string", "format": "uri", "nullable": true },
          "source_verified": { "type": "boolean", "description": "True when the source URL resolved at verification time." }
        }
      },
      "Insight": {
        "type": "object",
        "properties": {
          "insight_type": { "type": "string", "enum": ["industry_impact", "predicted_outcome", "precedent_analysis"] },
          "content":      { "type": "string" },
          "citations":    { "type": "array", "items": { "type": "object", "properties": {
            "kind": { "type": "string", "enum": ["lawsuit", "regulation"] }, "id": { "type": "string", "format": "uuid" }, "name": { "type": "string" }
          } } },
          "confidence":   { "type": "number", "format": "float", "nullable": true },
          "generated_at": { "type": "string", "format": "date-time" }
        }
      },
      "Mention": {
        "type": "object",
        "properties": {
          "url":          { "type": "string", "format": "uri" },
          "host":         { "type": "string" },
          "title":        { "type": "string", "nullable": true },
          "author":       { "type": "string", "nullable": true },
          "publish_date": { "type": "string", "format": "date", "nullable": true },
          "excerpt":      { "type": "string", "nullable": true }
        }
      },
...
    }
  }
}